+++ src/lambda/write-code.js
@@ -6,9 +6,9 @@ let ws = require('./templates/ws')
 let scheduled = require('./templates/scheduled')
 let streams = require('./templates/streams')
 
-module.exports = function writeCode ({ handlerFile, type, runtime }, callback) {
+module.exports = function writeCode ({ handlerFile, type, runtime, body }, callback) {
   let types = { http, events, queues, ws, scheduled, streams }
   let run = runtime.split(/\d/)[0]
-  let body = types[type][run]
+  body = body || types[type][run]
   writeFile(handlerFile, body, callback)
 }